requirements.txt
streamlit==1.28.1
pytesseract==0.3.10
Pillow==10.0.1
python-docx==0.8.11
pandas==2.1.1
opencv-python==4.8.1.78
numpy==1.24.3
easyocr==1.7.0
torch==2.0.1
torchvision==0.15.2